move define_char! decl to grammar module
